rename layers_pt to layers
