+++ sockeye/train.py
@@ -36,7 +36,7 @@ from . import checkpoint_decoder
 from . import constants as C
 from . import data_io
 from . import encoder
-from . import layers_pt
+from . import layers
 from . import loss_pt
 from . import lr_scheduler
 from . import model_pt
